requirements.txt
python-dotenv 
PyPDF2
pydantic
pydantic_ai